refactor trace writer finalize chunk
